refactor: recognise these three failures structurally, not by message text
Each guard had a structured signal available that the message match stood in for:

- sorting by an unknown property throws with PathElementException in the cause
  chain, so the handler no longer reads Hibernate's prose at all. This is what the
  branch set out to fix, and it can no longer rot the same way.
- FrontendUrlProvider was asserting its own precondition by catching the
  IllegalStateException Spring throws for it. It now asks RequestContextHolder for
  ServletRequestAttributes directly.
- AzureFileStorageFactory compared the SDK's message for equality to tell a
  malformed connection string from any other client failure. Only connectionString()
  can report that, so the try now scopes just that call and any
  IllegalArgumentException from it means the same thing.

Sentry no longer receives the unknown-sort-property case, which answers 400 and is
not a server fault. ExceptionHandlersTest moves to io.tolgee.unit, where the
module's other non-Spring tests live.
