+++ backend/data/src/main/kotlin/io/tolgee/component/FrontendUrlProvider.kt
@@ -2,6 +2,8 @@ package io.tolgee.component
 
 import io.tolgee.configuration.tolgee.TolgeeProperties
 import org.springframework.stereotype.Component
+import org.springframework.web.context.request.RequestContextHolder
+import org.springframework.web.context.request.ServletRequestAttributes
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder
 
 @Component
@@ -19,20 +21,16 @@ class FrontendUrlProvider(
     }
 
   private fun getFromServerRequest(): String {
-    try {
-      val builder = ServletUriComponentsBuilder.fromCurrentRequestUri()
-      builder.replacePath("")
-      builder.replaceQuery("")
-      return builder.build().toUriString()
-    } catch (e: IllegalStateException) {
-      if (e.message?.contains("No current ServletRequestAttributes") == true) {
-        throw IllegalStateException(
-          "Trying to find frontend url, but there is no current request. " +
-            "You will have to specify frontend url in application properties.",
-        )
-      }
-      throw e
+    if (RequestContextHolder.getRequestAttributes() !is ServletRequestAttributes) {
+      throw IllegalStateException(
+        "Trying to find frontend url, but there is no current request. " +
+          "You will have to specify frontend url in application properties.",
+      )
     }
+    val builder = ServletUriComponentsBuilder.fromCurrentRequestUri()
+    builder.replacePath("")
+    builder.replaceQuery("")
+    return builder.build().toUriString()
   }
 
   fun getSubscriptionsUrl(organizationSlug: String): String =
